% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aggregate_rast_list}
\alias{aggregate_rast_list}
\title{Aggregates a raster or list of rasters using a vector of factors}
\usage{
aggregate_rast_list(fact, rlist)
}
\arguments{
\item{fact}{Vector of integers giving factors to aggregate by}

\item{rlist}{List of rasters to be aggregated}
}
\value{
List of rasters aggregated by specified factors, with original raster or list prepended to it
}
\description{
Aggregates a raster or list of rasters using a vector of factors
}

