% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.data.table.raster.R
\name{as.data.table.raster}
\alias{as.data.table.raster}
\title{Transform raster to data.table}
\usage{
as.data.table.raster(x, row.names = NULL, optional = FALSE, xy = FALSE,
  inmem = canProcessInMemory(x, 2), ...)
}
\arguments{
\item{x}{Raster* object}

\item{row.names}{`NULL` or character vector giving row names for data frame.
Missing values are not allowed}

\item{optional}{logical. If `TRUE`, setting row names and converting column
names (to syntactic names: see make.names) is optional}

\item{xy}{logical. If `TRUE`, also return the spatial coordinates}

\item{...}{Additional arguments (none) passed to `raster::as.data.frame`}

\item{centroids}{logical. If TRUE return the centroids instead of all spatial
coordinates (only relevant if xy=TRUE)}

\item{sepNA}{logical. If TRUE the parts of the spatial objects are separated
by lines that are NA (only if xy=TRUE and, for polygons, if centroids=FALSE}
}
\value{
returns a data.table object
}
\description{
Transform raster to data.table
}
\note{
This has been amended to fix the blocksize parameter, which was set to
2.
}
\examples{
logo <- brick(system.file("external/rlogo.grd", package="raster"))
v <- as.data.table(logo)
}
\author{
etiennebr, from https://gist.github.com/etiennebr/9515738/download#
}

