% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landcover-functions.R
\name{assignLCPct}
\alias{assignLCPct}
\title{Block processing function to assign new values to different layers in brick}
\usage{
assignLCPct(x, y, fname)
}
\arguments{
\item{x}{Brick with layers containing binary landcover classes}

\item{y}{Vector containing the values to assign to each layer in brick (thus one element per layer)}

\item{fname}{Output file name for reclassified brick}
}
\value{
Raster with altered values
}
\description{
Block processing function to assign new values to different layers in brick
}

