% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bias-stats.R
\name{bias_stats}
\alias{bias_stats}
\title{Calculate weighted bias statistics from raster}
\usage{
bias_stats(rbins, rwgts, rmask, reval, fun, stnm, evalnm, weighted, rnd = 2,
  trim_wgt = TRUE)
}
\arguments{
\item{rbins}{Raster of bins within which statistic should be calculated}

\item{rwgts}{Raster defining weighting of pixel towards calculation of stat}

\item{rmask}{Binary masking raster}

\item{reval}{Raster containing the bias statistics to evaluate}

\item{fun}{Function to evaluate--should be able to handle weights}

\item{stnm}{Name of statistic (for output column header)}

\item{evalnm}{Name of value encoded in reval's cells (for column header)}

\item{weighted}{TRUE or FALSE, whether statistic is weighted or not}

\item{rnd}{Rounding value to apply to statistic}

\item{trim_wgt}{Remove column of weights from output table}
}
\value{
A data.table of output statistics
}
\description{
Calculate weighted bias statistics from raster
}
\details{
The statistic of interest is calculated across i) the entire masked
in area of the raster, ii) the entire non-NA part of the raster (masked in
and masked-out), iii) within each bin in the masked in-area, iv) within each
bin for the masked and un-masked area. There are thus two output statistics
for each category (whole raster, and for each bin), the second of which is
indicated by the suffix ".nm", meaning it was calculated for all non-NA areas.
If an unweighted statistic is required, then for a mean value rwgts should be
a raster of zeros
}

