% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bias-stats.R
\name{bias_stats_list}
\alias{bias_stats_list}
\title{Calculate weighted bias statistics from list of rasters}
\usage{
bias_stats_list(rbinsl, rwgtsl, rmaskl, revall, fun, stnm, evalnm, weighted,
  rnd = 2, trim_wgt = TRUE, silent = FALSE)
}
\arguments{
\item{rbinsl}{1-level list of binning rasters}

\item{rwgtsl}{1-level list of rasters defining pixel weightings}

\item{rmaskl}{2-level list of binary masking rasters}

\item{revall}{2-level list of raster containing bias statistics to evaluate}

\item{fun}{Function to evaluate--should be able to handle weights}

\item{stnm}{Name of statistic (for output column header)}

\item{evalnm}{Name of value encoded in reval's cells (for column header)}

\item{weighted}{TRUE or FALSE, whether statistic is weighted or not}

\item{rnd}{Rounding value to apply to statistic}

\item{trim_wgt}{Remove column of weights from output table}
}
\value{
A list of data.table of output statistics
}
\description{
Calculate weighted bias statistics from list of rasters
}
\details{
This function runs the bias_stats function over a list of rasters
}

