% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias-statsw.R
\name{bias_statsw_list}
\alias{bias_statsw_list}
\title{Calculate weighted bias statistics from list of rasters}
\usage{
bias_statsw_list(refl, awgtsl, rerrorl, evalnm, fun, stnm, rnd = 2,
  rweight = TRUE, aweight = TRUE, trim_wgt = TRUE, silent = FALSE)
}
\arguments{
\item{refl}{1-level list of reference rasters}

\item{awgtsl}{1-level list of area weighting rasters}

\item{rerrorl}{2-level list of spatial error rasters}

\item{evalnm}{Name of value encoded in reval's cells (for column header)}

\item{fun}{Function to evaluate--should be able to handle weights}

\item{stnm}{Name of statistic (for output column header)}

\item{rnd}{Rounding value to apply to statistic}

\item{rweight}{Apply reference weighting or not}

\item{aweight}{Apply area weighting or not}

\item{trim_wgt}{Remove column of weights from output table}

\item{rnd}{Rounding value to apply to statistic}
}
\value{
A list of data.table output statistics
}
\description{
Calculate weighted bias statistics from list of rasters
}
\details{
This function runs the bias_statsw function over a list of rasters
}
