% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bin_stats}
\alias{bin_stats}
\title{Function to run various statistics over different lists produced by bin_value}
\usage{
bin_stats(stat.list, fun.list, bins)
}
\arguments{
\item{stat.list}{Output list created by bin_values function}

\item{fun.list}{List of functions, passed as e.g., list(mean, sd), or list(mean)}

\item{bins}{A vector defining the bin ranges, e.g. seq(0, 100, 10) for bins of 10%}
}
\value{
List of specified statistics applied per bin value
}
\description{
Function to run various statistics over different lists produced by bin_value
}
\details{
List of output tables, one table per level, in a list corresponding to different statistics
}

