% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bin_values}
\alias{bin_values}
\title{Groups the values of one raster into bins that correspond to ranges of values in another raster}
\usage{
bin_values(bin.rast, rast.to.bin, bins)
}
\arguments{
\item{bin.rast}{The "master raster", or that which has the values defining the bin ranges}

\item{rast.to.bin}{The raster whose values will be binned}

\item{bins}{A vector defining the bin ranges, e.g. seq(0, 100, 10) for bins of 10\%}
}
\value{
List of differences between rasters
}
\description{
Groups the values of one raster into bins that correspond to ranges of values in another raster
}
\details{
This requires the two inputs to be of the same extent in order for values to correspond
}

