% Generated by roxygen2 (4.0.1): do not edit by hand
\name{diff_rast_list}
\alias{diff_rast_list}
\title{Takes difference between single raster and each raster in list of rasters, organized by aggregation level}
\usage{
diff_rast_list(ind, subtractor, subtractee, whichdiff, abs = FALSE)
}
\arguments{
\item{ind}{Vector of integers used to index into first level of raster lists}

\item{subtractor}{Two level list of rasters, with first level being aggregation level}

\item{subtractee}{Raster/raster list to difference against, with same structure as subtractor}

\item{whichdiff}{Index or name determing which raster from rdifflist per factor level will be used to diff}

\item{abs}{TRUE or FALSE, for whether actual or absolute difference should be taken (default is FALSE)}
}
\description{
Takes difference between single raster and each raster in list of rasters, organized by aggregation level
}

