% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landcover-functions.R
\name{disaggregate_rast_list}
\alias{disaggregate_rast_list}
\title{Disaggregates a raster or list of rasters}
\usage{
disaggregate_rast_list(names1, names2, rlist, rmask)
}
\arguments{
\item{names1}{names or indices for outer loop of 2 level raster list}

\item{names2}{inner list names denoting aggregation level (see details)}

\item{rlist}{List of rasters to be aggregated}

\item{mask}{Masking raster to be applied}
}
\value{
List of rasters disaggregated back to original resolution
}
\description{
Disaggregates a raster or list of rasters
}
\details{
This is essentially the reverse of aggregate_rast_list, thus has a
structure that is highly specific to this project. The names 2 vector should
be of the form c("f1", "f5", "f10"), etc, where the f stands for factor and
the number gives the factor by which the raster being disaggregated was
aggregated from its original resolution.
}

