% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dt_to_raster.R
\name{dt_to_raster}
\alias{dt_to_raster}
\title{Transform data.table to raster*}
\usage{
dt_to_raster(dt, CRSobj, filename = NULL)
}
\arguments{
\item{dt}{data.table with x, y coordinates}

\item{CRSobj}{crs object}

\item{filename}{NULL, else raster* will be written to specified file}
}
\value{
returns a data.table object
}
\description{
Transform data.table to raster*
}

