% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_bin_values}
\alias{extract_bin_values}
\title{Applies bin_values to list}
\usage{
extract_bin_values(bin.rast, which.rast, rast.to.bin, bins)
}
\arguments{
\item{bin.rast}{The "master raster", or that which has the values defining the bin ranges}

\item{which.rast}{Index of raster in bin.rast level 2 on which to do the binning}

\item{rast.to.bin}{The raster whose values will be binned}

\item{bins}{A vector defining the bin ranges, e.g. seq(0, 100, 10) for bins of 10%}
}
\value{
List of values extracting from rasters corresponding to different bins
}
\description{
Applies bin_values to list
}
\details{
This requires the two inputs to be of the same extent in order for values to correspond
}

