% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lpfunc}
\alias{lpfunc}
\title{For over-plotting custom point or short horizontal lines}
\usage{
lpfunc(x, y, w = NULL, col, size = 1, type, pch = NULL)
}
\arguments{
\item{x}{x-coordinate for point or center of line}

\item{y}{single y-coordinate}

\item{w}{if line, specify total width it should be}

\item{col}{Color}

\item{size}{cex value for points, lwd for line}

\item{type}{"line" or "pt"}

\item{pch}{For points, e.g. "*", 20}
}
\value{
point or lines on plot
}
\description{
For over-plotting custom point or short horizontal lines
}
\keyword{internal}

