% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makeTransparent}
\alias{makeTransparent}
\title{Make any color transparent}
\usage{
makeTransparent(someColor, alpha = 100)
}
\arguments{
\item{someColor}{Color name, e.g. "red"}

\item{alpha}{Transparency value, from 0-255}
}
\value{
Transparent color value
}
\description{
Make any color transparent
}
\note{
Function found at 
stackoverflow.com/questions/8047668/transparent-equivalent-of-given-color
}
\keyword{internal}
