% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{polyfunc2}
\alias{polyfunc2}
\title{Create arbitrary polygon on plot}
\usage{
polyfunc2(x, y, w = 0.5, col, bcol, lwd = 1)
}
\arguments{
\item{x}{coordinate on which to center polygon on x-axis}

\item{y}{upper and lower coordinates for polygon on y-axis}

\item{w}{polygon width in x dimension}

\item{col}{fill color}

\item{bcol}{border color}

\item{lwd}{border width}
}
\value{
plotted polygon on previously called plot
}
\description{
Create arbitrary polygon on plot
}
\keyword{internal}

