% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{predfunc}
\alias{predfunc}
\title{GAM Prediction function}
\usage{
predfunc(model, newdat)
}
\arguments{
\item{model}{Model object}

\item{newdat}{data.frame havng dummy values for predictors}
}
\value{
Vector of predicted values with confidence intervals, for curve
plotting
}
\description{
GAM Prediction function
}
\keyword{internal}

