% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reshape_diff_list}
\alias{reshape_diff_list}
\title{Function to reshape list output created by extract_bin_values}
\usage{
reshape_diff_list(diff.list, sens.ind, level.vec, bin.len)
}
\arguments{
\item{diff.list}{Output list created by extract_bin_values function}

\item{sens.ind}{Index vector, named or numbered, for sensors (lowest level of list)}

\item{level.vec}{A vector defining the number of aggregation levels (second level of list)}

\item{bin.len}{The total number of bins used in extract_bin_values}
}
\value{
A reshaped list, by sensor, factor, then bin
}
\description{
Function to reshape list output created by extract_bin_values
}
\details{
This function assumes that the binned differences for each sensor were calculated from multiple
difference permutations, e.g. there should be three elements of the binned difference list, one created by
the 2011 GTI dataset, one by the mean of the 2007 and 2011 datasets, and one from the 2007 set.
}

