% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reshape_reshape_list}
\alias{reshape_reshape_list}
\title{Function to reshape list output created by reshape_diff_list, in this case to do one further pooling}
\usage{
reshape_reshape_list(resh.diff.list, sens.pat, level.vec, bin.len)
}
\arguments{
\item{resh.diff.list}{Output list created by reshape_diff_list}

\item{sens.pat}{Index vector of character patterns for the difference bins that need to be further pooled}

\item{level.vec}{A vector defining the number of aggregation levels (second level of list)}

\item{bin.len}{The total number of bins used in extract_bin_values}
}
\value{
A reshaped list, by sensor, factor, then bin
}
\description{
Function to reshape list output created by reshape_diff_list, in this case to do one further pooling
}
\details{
This function was created to do one further level of pooling of the binned differences resulting
from the MODIS and GlobCover high, medium, and low variants. It assumes that each level of the list is
named.
}

