% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{spathfunc}
\alias{spathfunc}
\title{Helps get paths to system files installed with package}
\usage{
spathfunc(x, y = "lc")
}
\arguments{
\item{x}{Name of system file to find}

\item{y}{Sub-directory under inst/extdata where file is stored (details)}
}
\value{
Full path to system file
}
\description{
Helps get paths to system files installed with package
}
\details{
For argument y, use "lc" for landcover, "yl" for "yield", "cl" 
for "climate", "ab" for "abm", "cb" for "carbon", "et" for "et"
}
\keyword{internal}
