% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{termfunc}
\alias{termfunc}
\title{Predict values from single GAM model term}
\usage{
termfunc(model, newdat)
}
\arguments{
\item{model}{Model object}

\item{newdat}{data.frame havng dummy values for predictors}
}
\value{
Vector of predicted values with confidence intervals, for curve 
plotting
}
\description{
Predict values from single GAM model term
}
\keyword{internal}

