% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract_stat.R
\name{extract_stat}
\alias{extract_stat}
\title{Extract specific statistics from bias_stats data.table}
\usage{
extract_stat(ols, ils, bins, stat, val, dt, type = "density")
}
\arguments{
\item{ols}{vector with value(s) to subset from outer list (ol) variable}

\item{ils}{vector with value(s) to subset from inner list (il) variable}

\item{bins}{vector naming the statistical aggregation level to subset}

\item{stat}{statistic to subset}

\item{val}{Column header containing the value of interest (N or bias stat)}

\item{dt}{Name of the bias_stat produced data.table}

\item{type}{"density" for normal mode, "nodensity" for density independent}
}
\value{
Subset of statistics from bias_stat data.table
}
\description{
Extract specific statistics from bias_stats data.table
}
\details{
The type parameter allows for density independent and density
dependent statistics to be extracted. If density independent, the mean is
calculated as the mean of bin-wise means.
}

